Android: Fix android bearer implementation.
1. If the interface used by a configuration changes, remove the whole
configuration and replace it with a new one (instead of updating it).
While this isn't intuitive, this is the exact behavior we used to have
on android when using the generic plugin...

2. Setting the session state to roaming is not correct in this context
as it's a transitional state where we should end up in connected state
(forced roaming). Before this patch we would get stuck in the roaming
state for mobile networks.

4. Use QNetworkInterface::name() instead of
QNetworkInterface::humanReadableName(), as this is the expected
value when querying the interface used.

3. Don't re-iterate through the interface list for each configuration.

Task-number: QTBUG-41832
Change-Id: I315f725434bc6a1a8dca13dffd41f606c87bd06d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>