QMacStyle: More tweaks for editable QComboBox
Note that at one point we need to use Cocoa to render the combo box, but
only if we're dealing with Qt Quick controls.

Also worth noticing, there's currently a bug in Cocoa when rendering
inactive combob boxes. We faithfully reproduce it in Qt for now. We'll
fix it when Apple does.

Finally, we need to start constraininig the combo boxes height. Cocoa
has not supported variable height combo boxes for years, and will even
spit the following warning if we try to do something smart.

   This application is trying to draw a very large combo box, 28 points
   tall.  Vertically resizable combo boxes are not supported, but it
   happens that 10.4 and previous drew something that looked kind of sort
   of okay.  The art in 10.5 does not break up in a way that supports that
   drawing.  This application should be revised to stop using large combo
   boxes.  This warning will appear once per app launch.

Task-number: QTBUG-40833
Task-number: QTBUG-42067
Change-Id: I6512a6a581d446a28585db22fe4dbeac09499321
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>